+++ header.php
@@ -121,13 +121,7 @@
 			<!-- Multisite: Show site name below logo -->
 			<a href="<?php echo esc_url(home_url('/')); ?>"><img src="<?php echo LOOPIS_THEME_URI; ?>/assets/img/LOOPIS_logo.png" alt="LOOPIS-logo" id="header-logo"></a>
 			<div class="header-area">📍<?php echo get_bloginfo('name'); ?></div>
-
-			<!-- Profile/main site link -->
-			<?php if (is_user_logged_in()) : ?>
-				<div class="header-faq" onclick="location.href='<?php echo esc_url( network_home_url('/user/') ); ?>'">👤</div>
-			<?php else: ?>
-				<div class="header-faq" onclick="location.href='<?php echo esc_url( network_home_url() ); ?>'">🗺</div>
-			<?php endif; ?>
+			<div class="header-faq" onclick="location.href='<?php echo esc_url( network_home_url() ); ?>'">🗺</div>
 			</div>
 		</header>