+++ templates/links/log-in-button.php
@@ -8,4 +8,4 @@ if (!defined('ABSPATH')) {
 }
 ?>
 
-<p><button type="button" class="green" onclick="window.location.href='<?php echo esc_url(get_login_url()); ?>'">Logga in</button></p>
+<p><button type="button" class="green" onclick="window.location.href='<?php echo esc_url(get_loopis_login_url()); ?>'">Logga in</button></p>