+++ templates/links/log-in.php
@@ -8,4 +8,4 @@ if (!defined('ABSPATH')) {
 }
 ?>
 
-<a href="<?php echo esc_url(get_login_url()); ?>">👤 Logga in</a>
+<a href="<?php echo esc_url(get_loopis_login_url()); ?>">👤 Logga in</a>